Init repo and upload work so far
I have successfully coded a Godot project that puts the main window between the user's background and icons on the windows desktop.  There is some functionality missing which I will add in the near future.

**WARNING**
The game should not be built with this functionality as there is no way to close the window without either using task manager or exiting the debugging mode in editor. I REPEAT DO NOT BUILD YOUR GAME WITH THIS.

I will add a little X button later that will allow for it to be closed but for now it won't be.

TODO:
Figure out how to hook into the window close event from Godot and call clearDesktop() when doing so.

Add a button to close the window so that the project can be run standalone without editor.

PUT YOUR RESEARCH IN THE COMMENTS OF THIS SO THAT YOU GIVE CREDIT SMART GUY