Added my research resources.
Disclaimer, a majority of this is copying others code as there are some really nicely done reference works out there for what Samith wanted to do. I got everything set up multiple times with various methods but the current one is the one that works, and is extremely reliable.

Pros: It renders and only calls it's setup for background once, never needing to check it again to ensure the user hasn't moved it upwards in the windows layering.

Cons:
It's window border goes back to being the windows 7 border as we are technically embedding a window within a window, I believe I can fix it given enough time but just use full screen mode instead please.

TODO:
Hook Window closing so we can clear the desktop and redraw the background properly.

Add an X in the top right of the screen so the user can close the window after it is opened without needing to use task manager or the stop button in the editor.